Use TSCC's jsFiles param to pull in a Soy Template
This CL adds support for using Closure (Soy) Templates from Typescript.
TSCC has a configuration param 'jsFiles' which should be a list of
Closure Javascript files that will be passed to the Closure Compiler
(https://github.com/theseanl/tscc/issues/574).

The Closure Template Compiler has two modes: it can be used to generate
plain Javascript or Closure Javascript (see the README in the zip file
of a release of the Closure Template on Github). I tried using the the
plain Javascript version to no avail. The Closure Javascript version
ended up being the one that worked. However, there were some
complications.

The Closure Javascript version of the template depends on (a) some
Closure Javascript in the Closure Templates repo and (b) the Closure
Library. These need to also be passed in via jsFiles (except for the
library's base.js and reflect.js to avoid 'duplicate provides' errors - it
seems TSCC already includes vendored versions of these in compilation).

I tried
using the google-closure-library with the template generated by a Bazel
closure_js_template_library rule, but IIRC there were some mismatches between
the the API in google-closure-library and the generated Closure template.

In the end, I ended up pulling in the Closure Template Compiler (into
third_party/closure-templates) and the Closure Library
(third_party/closure-library) as git submodules. The
google-closure-library NPM package might have worked; however, I am pretty sure
it is necessary to use the _Template Compiler_ at HEAD instead of the Bazel
rule's version. This is because the Bazel rule appears to bring in and
use an old version of soyutils_usegoog.js that's incompatible with the
soy package in the Closure Library (see rules_closure's soy_jssrc.BUILD
and the WORKSPACE dependency com_google_template_soy_jssrc).

Also, the Closure Template Compiler was broken at HEAD, so I had to fix
that.